Add 1-string_nconcat to concatenate strings
This function will concatenate n bytes of char *s2 into a copy of
s1 and return a pointer to the new and longer string.
There was previously a function 1-strncat that seem to do the same.
But _strncat copies to an empty, fixed size string, while string_nconcat
creates a new dynamic size string